@@ -4,16 +4,18 @@
 	"description": "Display images in iTerm",
 	"license": "MIT",
 	"repository": "sindresorhus/term-img-cli",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
 	"bin": {
-		"term-img": "cli.js"
+		"term-img": "./cli.js"
 	},
 	"engines": {
-		"node": ">=8"
+		"node": ">=12"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -38,13 +40,13 @@
 		"gifs"
 	],
 	"dependencies": {
-		"get-stdin": "^6.0.0",
-		"meow": "^5.0.0",
-		"term-img": "^4.0.0"
+		"get-stdin": "^9.0.0",
+		"meow": "^9.0.0",
+		"term-img": "^6.0.0"
 	},
 	"devDependencies": {
-		"ava": "^1.0.1",
-		"execa": "^1.0.0",
-		"xo": "^0.23.0"
+		"ava": "^3.15.0",
+		"execa": "^5.0.0",
+		"xo": "^0.39.1"
 	}
 }
